requirements.txt
streamlit>=1.35.0
google-genai>=0.8.0
python-dotenv>=1.0.0